Vehicle Changes - June 13

+++ vehicle_update.php
@@ -158,6 +158,7 @@ $conn = mysqli_connect("localhost","fascalab_2020","w]zYV6X9{*BN","fascalab_2020
 $username1 = $_SESSION['username'];
 //input check value
 $checked = $_POST['check'];
+echo $checked;
 /* Requset vr_count */
 $vr_c = $_POST['vr_c'];
 
@@ -191,8 +192,8 @@ else if($checked=='Day/s'){
 
 }
 else{
-  $type1="Drop Off";
-  //echo '<div class=""><div class="panel-heading " style = "background-color:Red"> <p style = "color:white;font-size:16px;"> Type is required.  </p> </div></div>  '; 
+ // $type1="";
+ 
 }
 
 /* echo $type;
@@ -229,7 +230,12 @@ if ($conn->connect_error) {
    die("Connection failed: " . $conn->connect_error);
 }
 
+if($checked==''){
+ echo '<div class=""><div class="panel-heading " style = "background-color:Red"> <p style = "color:white;font-size:16px;"> Type is a required field.  </p> </div></div>  '; 
 
+}
+else
+{
   $query2 = mysqli_query($conn, "DELETE from vr_passengers where vrid='$vrno' ");
   /* echo "DELETE from vr_passengers where vrid='$vrno' ";
   exit(); */
@@ -270,6 +276,7 @@ else{
   echo '<div class=""><div class="panel-heading " style = "background-color:Red"> <p style = "color:white;font-size:16px;"> Error. </p> </div></div>  '; 
    
 }
+}
 
 }
 
@@ -283,7 +290,7 @@ else{
             <h1 align="">Update Vehicle Request</h1>
             
         <br>
-      <li class="btn btn-success"><a href="VehicleRequest.php" style="color:white;text-decoration: none;">Back</a></li>
+      <li class="btn btn-warning"><a href="VehicleRequest.php" style="color:white;text-decoration: none;">Back</a></li>
       <br>
       <br>
      
@@ -294,7 +301,7 @@ else{
         <form method="POST" action='' enctype="multipart/form-data" >
                 <table class="table"> 
               
-                <input hidden  class="" type="text" class="" style="height: 35px;" id="check" name="check" placeholder="check" >
+                <input hidden  class="" type="text" class="" style="height: 35px;" id="check" name="check" placeholder="check"  value ="<?php echo $type?>">
                 <input  hidden class="" type="text" class="" style="height: 35px;" id="vr_c" name="vr_c" placeholder="" value ="<?php echo $vrcount11?>">
                 
                 
@@ -665,7 +672,7 @@ else{
                         }
                         ?>
                         
-                        <textarea rows = "50" cols="1" name="passengers" id="passengers" style="text-align:left; border:none; border-bottom:1px solid black; font-size:15px;  height: 135px; width:100%;"><?php echo passengers($connect)?></textarea>
+                        <textarea required rows = "50" cols="1" name="passengers" id="passengers" style="text-align:left; border:none; border-bottom:1px solid black; font-size:15px;  height: 135px; width:100%;"><?php echo passengers($connect)?></textarea>
                         
                       </td>
                       <!-- Passengers/: -->
